FBCSR reference kernel testing now depends mostly on comparison with CSR kernels
This enabled reduction in FBCSR sample matrix code and removal of the
cpp file, so the test_fbcsr_sample library is no longer needed.
