+++ reference/test/matrix/CMakeLists.txt
@@ -3,7 +3,7 @@ ginkgo_create_test(csr_kernels)
 ginkgo_create_test(dense_kernels)
 ginkgo_create_test(diagonal_kernels)
 ginkgo_create_test(ell_kernels)
-ginkgo_create_test(fbcsr_kernels test_fbcsr_sample)
+ginkgo_create_test(fbcsr_kernels)
 ginkgo_create_test(hybrid_kernels)
 ginkgo_create_test(identity)
 ginkgo_create_test(permutation)
